requirements.txt
numpy==1.19.5
Pillow==8.4.0
protobuf==3.15.8
six==1.16.0
tensorboardX==2.4
torch==1.9.0+cu102
torchvision==0.10.0+cu102
tqdm==4.62.3
